// src/app/models/Types.tsx
export interface ScheduleEvent {
    Flightnr: string;
    Date: string;
    "Aircraft Type": string;
    Tail: string;
    Departure: string;
    Destination: string;
    Time_Depart: string;
    Time_Arrive: string;
    DutyID: string;
    DutyCode: string;
    Captain: string;
    "First Officer": string;
    "Flight Attendant": string;
}